// src/app/models/user.ts
export interface User {
    id: number;
    name: string;
    mail: string;
    rol: string;
    departament: string;
    management: string;
    phone: string;
    avatar: string;
    email?: string;
    idempleado: string;
    lastname: string;
    status: string;
    stringValue?: string;
    fecha?: Number;
    apellidoMaterno:string;
    apellidoPaterno:string;
    tipo:string;
    julio?:string;
    mapValue?:string;
}